requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.9.0
requests>=2.26.0
pyyaml>=6.0
python-dotenv>=0.19.1
signalwire-pom>=0.1.0
